Dev backend

+++ backend/orm/common/index.py
@@ -1,6 +1,7 @@
+from dependencies.authorization_user import has_authorization_object_single, has_authorization_object_collection
+from sqlalchemy import or_
 from sqlalchemy.exc import SQLAlchemyError
-from sqlalchemy.orm import joinedload
-from typing import Any, List
+from typing import Any
 from fastapi import HTTPException
 from sqlalchemy.orm import Session
 from schemas.common.pagination import MetadataSchema, PaginationSchema
@@ -17,21 +18,64 @@ def get_by_key_value(db: Session, model: Any, key: str, value):
     return db_object
 
 
-def get_by_id(db: Session, model: Any, id: int, relations: List[str] = []):
-    query = db.query(model)
-    for relation in relations:
-        query = query.options(joinedload(getattr(model, relation)))
-    db_object = query.filter(model.id == id).first()
+def user_autenthicated(token: str, db: Session):
+    from dependencies.authenticated_user import get_authenticated_user
+    return get_authenticated_user(token, db)
+
+
+async def get_by_id(
+    db: Session,
+    model: Any,
+    id: int,
+    token: str,
+    model_has_user_key: Any,
+):
+
+    db_object = db.query(model).filter(model.id == id).first()
+
     if db_object:
+        if (not await has_authorization_object_single(model, db, db_object, token, model_has_user_key)):
+            raise HTTPException(status.HTTP_401_UNAUTHORIZED)
+
         return db_object
     raise HTTPException(status.HTTP_404_NOT_FOUND)
 
 
-def get_all(db: Session, model: Any, common: PaginationSchema):
-    db_objects = db.query(model).offset(common.offset).limit(common.limit)
-    total = db.query(model).count()
+async def get_all(
+    db: Session,
+    model: Any,
+    common: PaginationSchema,
+    token: str,
+    filters: Any = None,
+    search_fields: list[str] = [],
+    allow_any: bool = False
+):
+
+    if (allow_any == False):
+        if (not await has_authorization_object_collection(db, token)):
+            raise HTTPException(status.HTTP_401_UNAUTHORIZED)
+
+    query = db.query(model)
+
+    if filters:
+        for attr, value in filters.__dict__.items():
+            if value is not None:
+                query = query.filter(getattr(model, attr) == value)
+
+    if common.q and search_fields:
+        search_query = or_(
+            *[getattr(model, field).ilike(f"%{common.q}%") for field in search_fields if hasattr(model, field)])
+        query = query.filter(search_query)
+
+    db_objects = query.offset(common.offset).limit(common.limit)
+    total = query.count()
     metadata = MetadataSchema(
-        count=db_objects.count(), total=total, offset=common.offset, limit=common.limit)
+        count=db_objects.count(),
+        total=total,
+        offset=common.offset,
+        limit=common.limit,
+        search_fields=search_fields
+    )
 
     return db_objects.all(), metadata
 
@@ -48,10 +92,19 @@ def create_object(db: Session, model: Any, schema: Any):
     return db_object
 
 
-def delete_object(db: Session, model: Any, id: int):
+async def delete_object(
+    db: Session,
+    model: Any,
+    id: int,
+    token: str = "",
+    model_has_user_key: Any = None,
+):
     db_object = db.query(model).filter(model.id == id).first()
     if not db_object:
         raise HTTPException(status.HTTP_404_NOT_FOUND)
+
+    if (token and not await has_authorization_object_single(model, db, db_object, token, model_has_user_key)):
+        raise HTTPException(status.HTTP_401_UNAUTHORIZED)
 
     try:
         db.delete(db_object)
@@ -62,18 +115,29 @@ def delete_object(db: Session, model: Any, id: int):
         raise HTTPException(status.HTTP_500_INTERNAL_SERVER_ERROR)
 
 
-def update_object(db: Session, model: Any, id: int, data: Any):
+async def update_object(
+    db: Session,
+    model: Any,
+    id: int,
+    data: Any,
+    token: str = "",
+    model_has_user_key: Any = None,
+):
+
     db_object = db.query(model).filter(model.id == id).first()
     if not db_object:
         raise HTTPException(status.HTTP_404_NOT_FOUND)
 
+    if (token and not await has_authorization_object_single(model, db, db_object, token, model_has_user_key)):
+        raise HTTPException(status.HTTP_401_UNAUTHORIZED)
+
     try:
         with db.begin_nested():
-            for key, value in data.dict().items():
+            for key, value in data.model_dump().items():
                 if hasattr(db_object, key):
                     setattr(db_object, key, value)
-            db.commit()
-            db.refresh(db_object)
+        db.commit()
+        db.refresh(db_object)
     except SQLAlchemyError:
         db.rollback()
         raise HTTPException(status.HTTP_500_INTERNAL_SERVER_ERROR)
